Moved all non-test classes from "test" package to "main" package
