+++ src/main/java/com/softserve/edu/greencity/ui/assertions/EcoNewsSuggestionsAssertion.java
