--- src/main/java/com/softserve/edu/greencity/ui/tools/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Package.
- */
-package com.softserve.edu.greencity.ui.tools;
