+++ src/main/java/com/softserve/edu/greencity/ui/tools/unused/Randomizer.java
@@ -1,4 +1,4 @@
-package com.softserve.edu.greencity.ui.tools;
+package com.softserve.edu.greencity.ui.tools.unused;
 
 import java.util.Random;
