+++ src/main/java/com/softserve/edu/greencity/ui/tools/unused/TabsHandler.java
@@ -1,4 +1,4 @@
-package com.softserve.edu.greencity.ui.tools;
+package com.softserve.edu.greencity.ui.tools.unused;
 
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
