+++ src/test/java/com/softserve/edu/greencity/ui/tests/createnews/CreateNewsPositiveTest.java
@@ -10,8 +10,8 @@ import com.softserve.edu.greencity.ui.pages.econews.EcoNewsPage;
 import com.softserve.edu.greencity.ui.pages.econews.SingleNewsPage;
 import com.softserve.edu.greencity.ui.pages.econews.TagsComponent;
 import com.softserve.edu.greencity.ui.tests.runner.GreenCityTestRunner;
-import com.softserve.edu.greencity.ui.tests.runner.LocalOnly;
-import com.softserve.edu.greencity.ui.tests.runner.RemoteSkipTestAnalyzer;
+import com.softserve.edu.greencity.ui.tools.testng.LocalOnly;
+import com.softserve.edu.greencity.ui.tools.testng.RemoteSkipTestAnalyzer;
 import com.softserve.edu.greencity.ui.tools.jdbc.services.EcoNewsService;
 import io.qameta.allure.Description;
 import org.testng.Assert;
