+++ src/test/java/com/softserve/edu/greencity/ui/tests/viewallnews/EcoNewsGridViewTest.java
@@ -7,8 +7,8 @@ import com.softserve.edu.greencity.data.econews.NewsDataRepository;
 import com.softserve.edu.greencity.ui.pages.econews.EcoNewsPage;
 import com.softserve.edu.greencity.ui.pages.econews.ItemComponent;
 import com.softserve.edu.greencity.ui.tests.runner.GreenCityTestRunner;
-import com.softserve.edu.greencity.ui.tests.runner.LocalOnly;
-import com.softserve.edu.greencity.ui.tests.runner.RetryAnalyzerImpl;
+import com.softserve.edu.greencity.ui.tools.testng.LocalOnly;
+import com.softserve.edu.greencity.ui.tools.testng.RetryAnalyzerImpl;
 import com.softserve.edu.greencity.ui.tools.jdbc.services.EcoNewsService;
 import io.qameta.allure.Description;
 import org.openqa.selenium.WebElement;
